backend-ai/requirements.txt
fastapi[standard]
uvicorn[standard]
sentence-transformers
numpy
motor
python-dotenv
